Refactor Chewy::Index::Observe
